Move utility interrupt functions and other utilities to appropriate modules
