+++ src/arch/x86_64/interrupts/mod.rs
@@ -4,11 +4,14 @@ use x86_64::structures::idt::{Idt, ExceptionStackFrame};
 use spin::Once;
 use device::pic::PICS;
 use device::keyboard::read_char;
-use utils::disable_interrupts_and_then;
+
 
 mod gdt;
 mod exceptions;
 mod irq;
+pub mod utils;
+
+pub use self::utils::*;
 
 const DOUBLE_FAULT_IST_INDEX: usize = 0;
