+++ src/arch/x86_64/memory/heap_allocator.rs
@@ -1,6 +1,6 @@
 use alloc::allocator::{Alloc, AllocErr, Layout};
 use linked_list_allocator::LockedHeap;
-use utils::disable_interrupts_and_then;
+use arch::interrupts::disable_interrupts_and_then;
 
 pub const HEAP_START: usize = 0o_000_001_000_000_0000;
 pub const HEAP_SIZE: usize = 500 * 1024;
